fix orderby type
