+++ products/logs/frontend/LogsScene.tsx
@@ -1,4 +1,4 @@
-import { useActions, useValues } from 'kea'
+import { useValues } from 'kea'
 
 import { IconGear } from '@posthog/icons'
 import { LemonBanner, LemonButton } from '@posthog/lemon-ui'
@@ -32,21 +32,8 @@ export function LogsScene(): JSX.Element {
 }
 
 const LogsSceneContent = (): JSX.Element => {
-    const {
-        tabId,
-        parsedLogs,
-        logsLoading,
-        totalLogsMatchingFilters,
-        sparklineLoading,
-        hasMoreLogsToLoad,
-        orderBy,
-        sparklineData,
-        sparklineBreakdownBy,
-        maxExportableLogs,
-    } = useValues(logsSceneLogic)
+    const { tabId } = useValues(logsSceneLogic)
     const { hasLogs, teamHasLogsCheckFailed } = useValues(logsIngestionLogic)
-    const { runQuery, fetchNextLogsPage, setOrderBy, addFilter, setDateRange, setSparklineBreakdownBy, zoomDateRange } =
-        useActions(logsSceneLogic)
     const openLogsSettings = useOpenLogsSettingsPanel()
 
     return (
@@ -85,25 +72,7 @@ const LogsSceneContent = (): JSX.Element => {
             )}
             <LogsSetupPrompt>
                 <div className="flex flex-col gap-2 py-2 h-[calc(100vh_-_var(--breadcrumbs-height-compact,_0px)_-_var(--scene-title-section-height,_0px)_-_5px_+_10rem)]">
-                    <LogsViewer
-                        tabId={tabId}
-                        logs={parsedLogs}
-                        loading={logsLoading}
-                        totalLogsCount={sparklineLoading ? undefined : totalLogsMatchingFilters}
-                        hasMoreLogsToLoad={hasMoreLogsToLoad}
-                        orderBy={orderBy}
-                        onChangeOrderBy={setOrderBy}
-                        onRefresh={runQuery}
-                        onLoadMore={fetchNextLogsPage}
-                        onAddFilter={addFilter}
-                        sparklineData={sparklineData}
-                        sparklineLoading={sparklineLoading}
-                        onDateRangeChange={setDateRange}
-                        sparklineBreakdownBy={sparklineBreakdownBy}
-                        onSparklineBreakdownByChange={setSparklineBreakdownBy}
-                        onExpandTimeRange={() => zoomDateRange(2)}
-                        maxExportableLogs={maxExportableLogs}
-                    />
+                    <LogsViewer id={tabId} />
                 </div>
             </LogsSetupPrompt>
         </>
