+++ products/logs/frontend/components/LogsViewer/LogsViewer.tsx
@@ -5,19 +5,18 @@ import { TZLabelProps } from 'lib/components/TZLabel'
 import { useKeyboardHotkeys } from 'lib/hooks/useKeyboardHotkeys'
 
 import { SceneDivider } from '~/layout/scenes/components/SceneDivider'
-import { DateRange, LogsSparklineBreakdownBy } from '~/queries/schema/schema-general'
-import { PropertyFilterType, PropertyOperator } from '~/types'
 
 import { LogsFilterBar } from 'products/logs/frontend/components/LogsViewer/Filters/LogsFilterBar/LogsFilterBar'
 import { logsViewerFiltersLogic } from 'products/logs/frontend/components/LogsViewer/Filters/logsViewerFiltersLogic'
 import { logsViewerConfigLogic } from 'products/logs/frontend/components/LogsViewer/config/logsViewerConfigLogic'
+import { logsViewerDataLogic } from 'products/logs/frontend/components/LogsViewer/data/logsViewerDataLogic'
+import { logsExportLogic } from 'products/logs/frontend/components/LogsViewer/logsExportLogic'
 import { VirtualizedLogsList } from 'products/logs/frontend/components/VirtualizedLogsList/VirtualizedLogsList'
 import { virtualizedLogsListLogic } from 'products/logs/frontend/components/VirtualizedLogsList/virtualizedLogsListLogic'
-import { LogsOrderBy, ParsedLogMessage } from 'products/logs/frontend/types'
 
 import { LogDetailsModal } from './LogDetailsModal'
 import { logDetailsModalLogic } from './LogDetailsModal/logDetailsModalLogic'
-import { LogsSparkline, LogsSparklineData } from './LogsViewerSparkline'
+import { LogsSparkline } from './LogsViewerSparkline'
 import { LogsViewerToolbar } from './LogsViewerToolbar'
 import { logsViewerLogic } from './logsViewerLogic'
 
@@ -25,65 +24,20 @@ const SCROLL_INTERVAL_MS = 16 // ~60fps
 const SCROLL_AMOUNT_PX = 8
 
 export interface LogsViewerProps {
-    tabId: string
-    logs: ParsedLogMessage[]
-    loading: boolean
-    totalLogsCount?: number
-    hasMoreLogsToLoad?: boolean
-    orderBy: LogsOrderBy
-    onChangeOrderBy: (orderBy: LogsOrderBy, source: 'header' | 'toolbar') => void
-    onRefresh?: () => void
-    onLoadMore?: () => void
-    onAddFilter?: (key: string, value: string, operator?: PropertyOperator, type?: PropertyFilterType) => void
-    sparklineData: LogsSparklineData
-    sparklineLoading: boolean
-    onDateRangeChange: (dateRange: DateRange) => void
-    sparklineBreakdownBy: LogsSparklineBreakdownBy
-    onSparklineBreakdownByChange: (breakdownBy: LogsSparklineBreakdownBy) => void
-    onExpandTimeRange?: () => void
-    maxExportableLogs: number
+    id: string
 }
 
-export function LogsViewer({
-    tabId,
-    logs,
-    loading,
-    totalLogsCount,
-    hasMoreLogsToLoad,
-    orderBy,
-    onChangeOrderBy,
-    onRefresh,
-    onLoadMore,
-    onAddFilter,
-    sparklineData,
-    sparklineLoading,
-    onDateRangeChange,
-    sparklineBreakdownBy,
-    onSparklineBreakdownByChange,
-    onExpandTimeRange,
-    maxExportableLogs,
-}: LogsViewerProps): JSX.Element {
+export function LogsViewer({ id }: LogsViewerProps): JSX.Element {
     return (
-        <BindLogic logic={logsViewerFiltersLogic} props={{ id: tabId }}>
-            <BindLogic logic={logsViewerConfigLogic} props={{ id: tabId }}>
-                <BindLogic logic={logDetailsModalLogic} props={{ tabId }}>
-                    <BindLogic logic={logsViewerLogic} props={{ tabId, logs, orderBy, onAddFilter }}>
-                        <LogsViewerContent
-                            loading={loading}
-                            totalLogsCount={totalLogsCount}
-                            hasMoreLogsToLoad={hasMoreLogsToLoad}
-                            orderBy={orderBy}
-                            onChangeOrderBy={onChangeOrderBy}
-                            onRefresh={onRefresh}
-                            onLoadMore={onLoadMore}
-                            sparklineData={sparklineData}
-                            sparklineLoading={sparklineLoading}
-                            onDateRangeChange={onDateRangeChange}
-                            sparklineBreakdownBy={sparklineBreakdownBy}
-                            onSparklineBreakdownByChange={onSparklineBreakdownByChange}
-                            onExpandTimeRange={onExpandTimeRange}
-                            maxExportableLogs={maxExportableLogs}
-                        />
+        <BindLogic logic={logsViewerFiltersLogic} props={{ id }}>
+            <BindLogic logic={logsViewerConfigLogic} props={{ id }}>
+                <BindLogic logic={logsViewerDataLogic} props={{ id }}>
+                    <BindLogic logic={logDetailsModalLogic} props={{ tabId: id }}>
+                        <BindLogic logic={logsViewerLogic} props={{ tabId: id }}>
+                            <BindLogic logic={logsExportLogic} props={{ tabId: id }}>
+                                <LogsViewerContent />
+                            </BindLogic>
+                        </BindLogic>
                     </BindLogic>
                 </BindLogic>
             </BindLogic>
@@ -91,39 +45,7 @@ export function LogsViewer({
     )
 }
 
-interface LogsViewerContentProps {
-    loading: boolean
-    totalLogsCount?: number
-    hasMoreLogsToLoad?: boolean
-    orderBy: LogsOrderBy
-    onChangeOrderBy: (orderBy: LogsOrderBy, source: 'header' | 'toolbar') => void
-    onRefresh?: () => void
-    onLoadMore?: () => void
-    sparklineData: LogsSparklineData
-    sparklineLoading: boolean
-    onDateRangeChange: (dateRange: DateRange) => void
-    sparklineBreakdownBy: LogsSparklineBreakdownBy
-    onSparklineBreakdownByChange: (breakdownBy: LogsSparklineBreakdownBy) => void
-    onExpandTimeRange?: () => void
-    maxExportableLogs: number
-}
-
-function LogsViewerContent({
-    loading,
-    totalLogsCount,
-    hasMoreLogsToLoad,
-    orderBy,
-    onChangeOrderBy,
-    onRefresh,
-    onLoadMore,
-    sparklineData,
-    sparklineLoading,
-    onDateRangeChange,
-    sparklineBreakdownBy,
-    onSparklineBreakdownByChange,
-    onExpandTimeRange,
-    maxExportableLogs,
-}: LogsViewerContentProps): JSX.Element {
+function LogsViewerContent(): JSX.Element {
     const {
         tabId,
         wrapBody,
@@ -132,7 +54,6 @@ function LogsViewerContent({
         isFocused,
         cursorIndex,
         cursorLogId,
-        logs,
         timezone,
         isSelectionActive,
         keyboardNavEnabled,
@@ -148,6 +69,12 @@ function LogsViewerContent({
         clearSelection,
         togglePrettifyLog,
     } = useActions(logsViewerLogic)
+    const { orderBy, sparklineBreakdownBy } = useValues(logsViewerConfigLogic)
+    const { setOrderBy, setSparklineBreakdownBy } = useActions(logsViewerConfigLogic)
+    const { logsLoading, parsedLogs, sparklineData, sparklineLoading, hasMoreLogsToLoad, totalLogsMatchingFilters } =
+        useValues(logsViewerDataLogic)
+    const { runQuery, fetchNextLogsPage } = useActions(logsViewerDataLogic)
+    const { setDateRange, zoomDateRange } = useActions(logsViewerFiltersLogic)
     const { cellScrollLefts } = useValues(virtualizedLogsListLogic({ tabId }))
     const { setCellScrollLeft } = useActions(virtualizedLogsListLogic({ tabId }))
     const messageScrollLeft = cellScrollLefts['message'] ?? 0
@@ -252,17 +179,17 @@ function LogsViewerContent({
             // arrowleft, arrowright, h, l handled by native keydown/keyup for smooth 60fps scrolling
             enter: {
                 action: () => {
-                    if (cursorIndex !== null && logs[cursorIndex]) {
-                        openLogDetails(logs[cursorIndex])
+                    if (cursorIndex !== null && parsedLogs[cursorIndex]) {
+                        openLogDetails(parsedLogs[cursorIndex])
                     }
                 },
                 disabled: !keyboardNavEnabled,
             },
             r: {
                 action: () => {
-                    if (onRefresh && !loading) {
+                    if (!logsLoading) {
                         resetCursor()
-                        onRefresh()
+                        runQuery()
                     }
                 },
                 disabled: !isFocused,
@@ -300,11 +227,11 @@ function LogsViewerContent({
             keyboardNavEnabled,
             isLogDetailsOpen,
             cursorIndex,
-            logs,
+            parsedLogs,
             openLogDetails,
             closeLogDetails,
-            onRefresh,
-            loading,
+            runQuery,
+            logsLoading,
             resetCursor,
             moveCursorDown,
             moveCursorUp,
@@ -321,17 +248,16 @@ function LogsViewerContent({
             <LogsSparkline
                 sparklineData={sparklineData}
                 sparklineLoading={sparklineLoading}
-                onDateRangeChange={onDateRangeChange}
+                onDateRangeChange={setDateRange}
                 displayTimezone={timezone}
                 breakdownBy={sparklineBreakdownBy}
-                onBreakdownByChange={onSparklineBreakdownByChange}
+                onBreakdownByChange={setSparklineBreakdownBy}
             />
             <SceneDivider />
             <LogsViewerToolbar
-                totalLogsCount={totalLogsCount}
+                totalLogsCount={sparklineLoading ? undefined : totalLogsMatchingFilters}
                 orderBy={orderBy}
-                onChangeOrderBy={(newOrderBy) => onChangeOrderBy(newOrderBy, 'toolbar')}
-                maxExportableLogs={maxExportableLogs}
+                onChangeOrderBy={(newOrderBy) => setOrderBy(newOrderBy, 'toolbar')}
             />
             {pinnedLogsArray.length > 0 && (
                 <VirtualizedLogsList
@@ -344,22 +270,22 @@ function LogsViewerContent({
                     disableInfiniteScroll
                     disableCursor
                     orderBy={orderBy}
-                    onChangeOrderBy={(newOrderBy) => onChangeOrderBy(newOrderBy, 'header')}
+                    onChangeOrderBy={(newOrderBy) => setOrderBy(newOrderBy, 'header')}
                 />
             )}
 
             <VirtualizedLogsList
-                dataSource={logs}
-                loading={loading}
+                dataSource={parsedLogs}
+                loading={logsLoading}
                 wrapBody={wrapBody}
                 prettifyJson={prettifyJson}
                 tzLabelFormat={tzLabelFormat}
                 showPinnedWithOpacity
                 hasMoreLogsToLoad={hasMoreLogsToLoad}
-                onLoadMore={onLoadMore}
-                onExpandTimeRange={onExpandTimeRange}
+                onLoadMore={fetchNextLogsPage}
+                onExpandTimeRange={() => zoomDateRange(2)}
                 orderBy={orderBy}
-                onChangeOrderBy={(newOrderBy) => onChangeOrderBy(newOrderBy, 'header')}
+                onChangeOrderBy={(newOrderBy) => setOrderBy(newOrderBy, 'header')}
             />
 
             <LogDetailsModal timezone={timezone} />
