+++ products/logs/frontend/components/LogsViewer/LogsViewerToolbar.tsx
@@ -16,14 +16,12 @@ export interface LogsViewerToolbarProps {
     totalLogsCount?: number
     orderBy: LogsOrderBy
     onChangeOrderBy: (orderBy: LogsOrderBy) => void
-    maxExportableLogs: number
 }
 
 export const LogsViewerToolbar = ({
     totalLogsCount,
     orderBy,
     onChangeOrderBy,
-    maxExportableLogs,
 }: LogsViewerToolbarProps): JSX.Element => {
     const { wrapBody, prettifyJson, timezone } = useValues(logsViewerLogic)
     const { setWrapBody, setPrettifyJson, setTimezone } = useActions(logsViewerLogic)
@@ -55,11 +53,7 @@ export const LogsViewerToolbar = ({
                     size="small"
                 />
                 <TimezoneSelect value={timezone} onChange={setTimezone} size="small" />
-                <LogsExportMenu
-                    totalLogsCount={totalLogsCount}
-                    orderBy={orderBy}
-                    maxExportableLogs={maxExportableLogs}
-                />
+                <LogsExportMenu totalLogsCount={totalLogsCount} />
             </div>
             <div className="flex items-center gap-4 flex-wrap">
                 {totalLogsCount !== undefined && totalLogsCount > 0 && (
