+++ products/logs/frontend/components/LogsViewer/config/logsViewerConfigLogic.ts
@@ -1,8 +1,10 @@
 import { actions, kea, key, path, props, reducers, selectors } from 'kea'
 
+import { LogsSparklineBreakdownBy } from '~/queries/schema/schema-general'
 import { FilterLogicalOperator } from '~/types'
 
 import { LogsViewerConfig, LogsViewerFilters } from 'products/logs/frontend/components/LogsViewer/config/types'
+import { LogsOrderBy } from 'products/logs/frontend/types'
 
 import type { logsViewerConfigLogicType } from './logsViewerConfigLogicType'
 
@@ -15,6 +17,10 @@ export const DEFAULT_LOGS_VIEWER_CONFIG: LogsViewerConfig = {
         filterGroup: { type: FilterLogicalOperator.And, values: [] },
     },
 }
+
+export const DEFAULT_SPARKLINE_BREAKDOWN_BY: LogsSparklineBreakdownBy = 'severity'
+
+export const DEFAULT_ORDER_BY: LogsOrderBy = 'latest'
 
 export interface LogsViewerConfigProps {
     id: string
@@ -31,6 +37,8 @@ export const logsViewerConfigLogic = kea<logsViewerConfigLogicType>([
             filter,
             value,
         }),
+        setSparklineBreakdownBy: (sparklineBreakdownBy: LogsSparklineBreakdownBy) => ({ sparklineBreakdownBy }),
+        setOrderBy: (orderBy: LogsOrderBy, source: 'header' | 'toolbar' = 'toolbar') => ({ orderBy, source }),
     }),
 
     reducers({
@@ -42,6 +50,19 @@ export const logsViewerConfigLogic = kea<logsViewerConfigLogicType>([
                     ...state,
                     [filter]: value,
                 }),
+            },
+        ],
+        sparklineBreakdownBy: [
+            DEFAULT_SPARKLINE_BREAKDOWN_BY as LogsSparklineBreakdownBy,
+            { persist: true },
+            {
+                setSparklineBreakdownBy: (_, { sparklineBreakdownBy }) => sparklineBreakdownBy,
+            },
+        ],
+        orderBy: [
+            DEFAULT_ORDER_BY as LogsOrderBy,
+            {
+                setOrderBy: (_, { orderBy }) => orderBy,
             },
         ],
     }),
