+++ products/logs/frontend/components/LogsViewer/logsExportLogic.ts
@@ -9,10 +9,13 @@ import { copyToClipboard } from 'lib/utils/copyToClipboard'
 import { sidePanelStateLogic } from '~/layout/navigation-3000/sidepanel/sidePanelStateLogic'
 import { PropertyGroupFilter, SidePanelTab } from '~/types'
 
-import { LogsOrderBy, ParsedLogMessage } from '../../types'
+import { logsViewerConfigLogic } from 'products/logs/frontend/components/LogsViewer/config/logsViewerConfigLogic'
+
+import { ParsedLogMessage } from '../../types'
 import { logsViewerFiltersLogic } from './Filters/logsViewerFiltersLogic'
+import { logsViewerDataLogic } from './data/logsViewerDataLogic'
 import type { logsExportLogicType } from './logsExportLogicType'
-import { LogsViewerLogicProps, logsViewerLogic } from './logsViewerLogic'
+import { logsViewerLogic } from './logsViewerLogic'
 
 function triggerBlobDownload(blob: Blob, filename: string): void {
     const url = URL.createObjectURL(blob)
@@ -40,19 +43,22 @@ export function getExportColumns(attributeColumns: string[]): string[] {
 
 export interface LogsExportLogicProps {
     tabId: string
-    orderBy: LogsOrderBy
 }
 
 export const logsExportLogic = kea<logsExportLogicType>([
     path((tabId) => ['products', 'logs', 'frontend', 'components', 'LogsViewer', 'logsExportLogic', tabId]),
     props({} as LogsExportLogicProps),
     key((props) => props.tabId),
-    connect((props: LogsExportLogicProps) => ({
+    connect(({ tabId }: LogsExportLogicProps) => ({
         values: [
-            logsViewerLogic({ tabId: props.tabId } as LogsViewerLogicProps),
+            logsViewerLogic({ tabId }),
             ['selectedLogsArray', 'attributeColumns'],
-            logsViewerFiltersLogic({ id: props.tabId }),
+            logsViewerFiltersLogic({ id: tabId }),
             ['filters', 'utcDateRange'],
+            logsViewerDataLogic({ id: tabId }),
+            ['maxExportableLogs'],
+            logsViewerConfigLogic({ id: tabId }),
+            ['orderBy'],
         ],
         actions: [sidePanelStateLogic, ['openSidePanel']],
     })),
@@ -63,7 +69,7 @@ export const logsExportLogic = kea<logsExportLogicType>([
         exportServerSide: (totalLogsCount?: number) => ({ totalLogsCount }),
     }),
 
-    listeners(({ actions, values, props }) => ({
+    listeners(({ actions, values }) => ({
         copySelectedLogs: () => {
             const selectedLogs = values.selectedLogsArray
             posthog.capture('logs bulk copy', { count: selectedLogs.length })
@@ -98,7 +104,7 @@ export const logsExportLogic = kea<logsExportLogicType>([
                 filterGroup: values.filters.filterGroup as PropertyGroupFilter,
                 severityLevels: values.filters.severityLevels,
                 serviceNames: values.filters.serviceNames,
-                orderBy: props.orderBy,
+                orderBy: values.orderBy,
             }
             posthog.capture('logs exported', { format: 'csv', source: 'server', totalLogsCount })
             try {
