+++ products/logs/frontend/components/LogsViewer/logsViewerLogic.test.ts
@@ -1,42 +1,75 @@
 import { expectLogic } from 'kea-test-utils'
 
+import { useMocks } from '~/mocks/jest'
+import { LogMessage } from '~/queries/schema/schema-general'
 import { initKeaTests } from '~/test/init'
 
 import { ParsedLogMessage } from 'products/logs/frontend/types'
 
+import { logsViewerFiltersLogic } from './Filters/logsViewerFiltersLogic'
+import { logsViewerConfigLogic } from './config/logsViewerConfigLogic'
+import { logsViewerDataLogic } from './data/logsViewerDataLogic'
 import { logsViewerLogic } from './logsViewerLogic'
 
+const createMockRawLog = (uuid: string): LogMessage => ({
+    uuid,
+    trace_id: 'trace-1',
+    span_id: 'span-1',
+    body: `Log ${uuid}`,
+    attributes: {},
+    timestamp: '2024-01-01T00:00:00Z',
+    observed_timestamp: '2024-01-01T00:00:00Z',
+    severity_text: 'info' as const,
+    severity_number: 9,
+    level: 'info' as const,
+    resource_attributes: {},
+    instrumentation_scope: 'test',
+    event_name: 'log',
+})
+
 const createMockParsedLog = (uuid: string): ParsedLogMessage => {
-    const baseLog = {
-        uuid,
-        trace_id: 'trace-1',
-        span_id: 'span-1',
-        body: `Log ${uuid}`,
-        attributes: {},
-        timestamp: '2024-01-01T00:00:00Z',
-        observed_timestamp: '2024-01-01T00:00:00Z',
-        severity_text: 'info' as const,
-        severity_number: 9,
-        level: 'info' as const,
-        resource_attributes: {},
-        instrumentation_scope: 'test',
-        event_name: 'log',
-    }
+    const baseLog = createMockRawLog(uuid)
     return {
         ...baseLog,
+        attributes: {},
         cleanBody: `Log ${uuid}`,
         parsedBody: null,
         originalLog: baseLog,
     }
 }
 
-const mockLogs = [createMockParsedLog('log-1'), createMockParsedLog('log-2'), createMockParsedLog('log-3')]
+const mockRawLogs = [createMockRawLog('log-1'), createMockRawLog('log-2'), createMockRawLog('log-3')]
+
+const TEST_ID = 'test-tab'
+
+function mountWithLogs(rawLogs: LogMessage[] = mockRawLogs): {
+    logic: ReturnType<typeof logsViewerLogic.build>
+    dataLogic: ReturnType<typeof logsViewerDataLogic.build>
+} {
+    const filtersLogic = logsViewerFiltersLogic({ id: TEST_ID })
+    filtersLogic.mount()
+    const configLogic = logsViewerConfigLogic({ id: TEST_ID })
+    configLogic.mount()
+    const dataLogic = logsViewerDataLogic({ id: TEST_ID, autoLoad: false })
+    dataLogic.mount()
+    if (rawLogs.length > 0) {
+        dataLogic.actions.setLogs(rawLogs)
+    }
+    const logic = logsViewerLogic({ tabId: TEST_ID })
+    logic.mount()
+    return { logic, dataLogic }
+}
 
 describe('logsViewerLogic', () => {
     let logic: ReturnType<typeof logsViewerLogic.build>
 
     beforeEach(() => {
-        // Clear localStorage to reset persisted state
+        useMocks({
+            post: {
+                '/api/environments/:team_id/logs/query/': () => [200, { results: [], maxExportableLogs: 5000 }],
+                '/api/environments/:team_id/logs/sparkline/': () => [200, []],
+            },
+        })
         localStorage.clear()
         initKeaTests()
     })
@@ -47,8 +80,7 @@ describe('logsViewerLogic', () => {
 
     describe('cursor navigation', () => {
         beforeEach(() => {
-            logic = logsViewerLogic({ tabId: 'test-tab', logs: mockLogs, orderBy: 'latest' })
-            logic.mount()
+            ;({ logic } = mountWithLogs())
         })
 
         it('sets cursor index', async () => {
@@ -212,8 +244,7 @@ describe('logsViewerLogic', () => {
 
     describe('empty logs', () => {
         beforeEach(() => {
-            logic = logsViewerLogic({ tabId: 'test-tab', logs: [], orderBy: 'latest' })
-            logic.mount()
+            ;({ logic } = mountWithLogs([]))
         })
 
         it('moveCursorDown does nothing when logs are empty', async () => {
@@ -238,8 +269,7 @@ describe('logsViewerLogic', () => {
         const mockLog2 = createMockParsedLog('log-2')
 
         beforeEach(() => {
-            logic = logsViewerLogic({ tabId: 'test-tab', logs: mockLogs, orderBy: 'latest' })
-            logic.mount()
+            ;({ logic } = mountWithLogs())
         })
 
         it('pins a log when not pinned', async () => {
@@ -298,8 +328,7 @@ describe('logsViewerLogic', () => {
 
     describe('focus state', () => {
         beforeEach(() => {
-            logic = logsViewerLogic({ tabId: 'test-tab', logs: mockLogs, orderBy: 'latest' })
-            logic.mount()
+            ;({ logic } = mountWithLogs())
         })
 
         it('defaults to not focused', () => {
@@ -323,8 +352,7 @@ describe('logsViewerLogic', () => {
 
     describe('display options', () => {
         beforeEach(() => {
-            logic = logsViewerLogic({ tabId: 'test-tab', logs: mockLogs, orderBy: 'latest' })
-            logic.mount()
+            ;({ logic } = mountWithLogs())
         })
 
         it('defaults wrapBody to true', () => {
@@ -354,8 +382,7 @@ describe('logsViewerLogic', () => {
 
     describe('timezone', () => {
         beforeEach(() => {
-            logic = logsViewerLogic({ tabId: 'test-tab', logs: mockLogs, orderBy: 'latest' })
-            logic.mount()
+            ;({ logic } = mountWithLogs())
         })
 
         it('defaults to UTC', () => {
@@ -384,8 +411,7 @@ describe('logsViewerLogic', () => {
 
     describe('multi-select', () => {
         beforeEach(() => {
-            logic = logsViewerLogic({ tabId: 'test-tab', logs: mockLogs, orderBy: 'latest' })
-            logic.mount()
+            ;({ logic } = mountWithLogs())
         })
 
         it('defaults to no selection', () => {
@@ -465,7 +491,7 @@ describe('logsViewerLogic', () => {
             })
 
             it('selects only provided logs when argument given', async () => {
-                const subset = [mockLogs[0], mockLogs[2]]
+                const subset = [logic.values.logs[0], logic.values.logs[2]]
                 await expectLogic(logic, () => {
                     logic.actions.selectAll(subset)
                 }).toFinishAllListeners()
@@ -542,7 +568,7 @@ describe('logsViewerLogic', () => {
                 expect(logic.values.selectedCount).toBe(2)
 
                 await expectLogic(logic, () => {
-                    logic.actions.setLogs([createMockParsedLog('new-log')])
+                    logic.actions.setLogs([createMockRawLog('new-log')])
                 }).toMatchValues({
                     selectedLogIds: {},
                     isSelectionActive: false,
@@ -553,8 +579,7 @@ describe('logsViewerLogic', () => {
 
     describe('attribute columns', () => {
         beforeEach(() => {
-            logic = logsViewerLogic({ tabId: 'test-tab', logs: mockLogs, orderBy: 'latest' })
-            logic.mount()
+            ;({ logic } = mountWithLogs())
         })
 
         describe('moveAttributeColumn', () => {
@@ -610,8 +635,7 @@ describe('logsViewerLogic', () => {
 
     describe('expansion', () => {
         beforeEach(() => {
-            logic = logsViewerLogic({ tabId: 'test-tab', logs: mockLogs, orderBy: 'latest' })
-            logic.mount()
+            ;({ logic } = mountWithLogs())
         })
 
         it('defaults to no expanded logs', () => {
@@ -657,8 +681,7 @@ describe('logsViewerLogic', () => {
 
     describe('per-row prettification', () => {
         beforeEach(() => {
-            logic = logsViewerLogic({ tabId: 'test-tab', logs: mockLogs, orderBy: 'latest' })
-            logic.mount()
+            ;({ logic } = mountWithLogs())
         })
 
         it('defaults to empty set', () => {
@@ -709,7 +732,7 @@ describe('logsViewerLogic', () => {
             expect(logic.values.prettifiedLogIds.size).toBe(2)
 
             await expectLogic(logic, () => {
-                logic.actions.setLogs([createMockParsedLog('new-log')])
+                logic.actions.setLogs([createMockRawLog('new-log')])
             }).toFinishAllListeners()
 
             expect(logic.values.prettifiedLogIds.size).toBe(0)
